# packages/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: KoalaCast REST API
  description: |
    Privacy-first podcast player API. Unix timestamps are explicitly identified as
    either seconds or milliseconds. Durations and playback positions use milliseconds.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
    description: Server root
security: []
...
paths:
...
  /api/v1/proxy/audio:
    parameters:
      - $ref: '#/components/parameters/ProxyURL'
    get:
      tags: [Proxy]
      operationId: getProxiedAudio
      summary: Stream external audio when the optional self-hosted relay is enabled
      responses:
        '200':
          description: Complete upstream audio
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          description: Partial audio response for a valid Range request
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: The self-hosted audio relay is disabled
        '413':
          description: Audio exceeds the 2 GiB relay limit
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/BadGateway'
    head:
      tags: [Proxy]
      operationId: headProxiedAudio
      summary: Fetch proxied audio headers when the optional relay is enabled
      responses:
        '200':
          description: Complete audio headers
        '206':
          description: Partial audio headers
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: The self-hosted audio relay is disabled
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/BadGateway'
  /api/v1/proxy/image:
    parameters:
      - $ref: '#/components/parameters/ProxyURL'
      - name: w
        in: query
        description: Target image width in pixels. Invalid values fall back to 300.
        schema:
          type: integer
          minimum: 1
          maximum: 1200
          default: 300
    get:
      tags: [Proxy]
      operationId: getProxiedImage
      summary: Fetch and resize an external image
      responses:
        '200':
          description: JPEG image, or a WebP fallback marked by X-KoalaCast-Image-Fallback
          headers:
            X-KoalaCast-Image-Fallback:
              schema:
                type: string
                enum: ['true']
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    head:
      tags: [Proxy]
      operationId: headProxiedImage
      summary: Fetch proxied image headers
      responses:
        '200':
          description: Image headers
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/proxy/chapters:
    parameters:
      - $ref: '#/components/parameters/ProxyURL'
    get:
      tags: [Proxy]
      operationId: getProxiedChapters
      summary: Fetch external Podcasting 2.0 chapters
      responses:
        '200':
          description: Publisher JSON, normalized when it contains recognized chapters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChaptersResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/BadGateway'
    head:
      tags: [Proxy]
      operationId: headProxiedChapters
      summary: Fetch proxied chapter headers
      responses:
        '200':
          description: Chapter response headers
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/BadGateway'
  /api/v1/proxy/transcript:
    parameters:
      - $ref: '#/components/parameters/ProxyURL'
    get:
      tags: [Proxy]
      operationId: getProxiedTranscript
      summary: Fetch and normalize an external transcript
      responses:
        '200':
          description: Publisher JSON or normalized SRT/VTT cues
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TranscriptCuesResponse'
                  - type: object
                    additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/BadGateway'
    head:
      tags: [Proxy]
      operationId: headProxiedTranscript
      summary: Fetch proxied transcript headers
      responses:
        '200':
          description: Transcript response headers
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/BadGateway'
  /api/v1/podcasts/discover:
    get:
      tags: [Podcasts]
      operationId: discoverPodcasts
      summary: Discover podcasts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: region
          in: query
          description: Two-letter storefront region.
          schema:
            type: string
        - name: languages
          in: query
          description: Comma-separated spoken-language codes.
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 60
        - name: include_explicit
          in: query
          description: Include podcasts explicitly marked by their publisher. Defaults to true for older clients.
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Discovery results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoverResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/podcasts/search:
    get:
      tags: [Podcasts]
      operationId: searchPodcasts
      summary: Search podcasts
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: languages
          in: query
          description: Comma-separated spoken-language codes.
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
        - name: country
          in: query
          deprecated: true
          description: Legacy alias for region.
          schema:
            type: string
        - name: include_explicit
          in: query
          description: Include podcasts explicitly marked by their publisher. Defaults to true for older clients.
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/podcasts/feed:
    post:
      tags: [Podcasts]
      operationId: addPodcastFeed
      summary: Resolve and ingest a podcast feed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFeedRequest'
      responses:
        '200':
          description: Canonical stored podcast
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Podcast'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/podcasts/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    get:
      tags: [Podcasts]
      operationId: getPodcast
      summary: Get a podcast
      responses:
        '200':
          description: Stored podcast
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Podcast'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/podcasts/{id}/episodes:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    get:
      tags: [Podcasts]
      operationId: listPodcastEpisodes
      summary: List podcast episodes
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: since
          in: query
          description: >
            Return only episodes whose publication timestamp is greater than or
            equal to this Unix-second value. Clients merge by episode id; the
            inclusive boundary prevents same-second publications from being lost.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Episode page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpisodesResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/episodes/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    get:
      tags: [Podcasts]
      operationId: getEpisode
      summary: Get an episode
      responses:
        '200':
          description: Stored episode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Episode'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/episodes/{id}/transcript:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    get:
      tags: [Podcasts]
      operationId: getEpisodeTranscript
      summary: Fetch a stored episode transcript
      parameters:
        - name: i
          in: query
          description: Zero-based transcript reference index.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Transcript content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/push/config:
    get:
      tags: [Push]
      operationId: getWebPushConfig
      summary: Read this server's Web Push capability
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: VAPID capability and public key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebPushConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/push/subscriptions:
    post:
      tags: [Push]
      operationId: registerWebPushSubscription
      summary: Register this browser's Push API subscription
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebPushSubscription'
      responses:
        '204':
          description: Subscription registered
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          description: Web Push is not configured on this server
    delete:
      tags: [Push]
      operationId: deleteWebPushSubscription
      summary: Remove this browser's Push API subscription
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint]
              additionalProperties: false
              properties:
                endpoint:
                  type: string
                  format: uri
      responses:
        '204':
          description: Subscription removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      operationId: registerUser
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Username is already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      operationId: loginWebSession
      summary: Log in with a browser session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded and koalacast_session cookie was set
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/device/login:
    post:
      tags: [Authentication]
      operationId: loginDevice
      summary: Issue native device credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceLoginRequest'
      responses:
        '200':
          description: Device credential issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceLoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/recovery/verify:
    post:
      tags: [Authentication]
      operationId: recoverAccount
      summary: Reset a password with a recovery code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoveryRequest'
      responses:
        '200':
          description: Password reset and replacement recovery code issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/status:
    get:
      tags: [Authentication]
      operationId: getAuthStatus
      summary: Get optional authentication status
      security:
        - {}
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Anonymous or authenticated status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatus'
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      operationId: getCurrentUser
      summary: Get the authenticated user
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserIdentity'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      summary: Revoke the current session or device credential
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/sessions:
    get:
      tags: [Authentication]
      operationId: listSessions
      summary: List active web sessions and device credentials
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Active sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/sessions/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    delete:
      tags: [Authentication]
      operationId: revokeSession
      summary: Revoke a web session or device credential
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/account:
    delete:
      tags: [Authentication]
      operationId: deleteAccount
      summary: Permanently delete the account and every record belonging to it
      description: >-
        Irreversible. Requires the account password or a recovery code in the
        body even though the request is already authenticated, so a borrowed
        unlocked device cannot delete an account. Every dependent row is removed
        in one transaction and all sessions and device credentials are revoked.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: At least one credential is required; clients may populate both fields.
              properties:
                password:
                  type: string
                recovery_code:
                  type: string
      responses:
        '204':
          description: Account deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/export:
    get:
      tags: [Authentication]
      operationId: exportAccount
      summary: Download every record stored for the account
      description: >-
        Returns account metadata, subscriptions, favorites and the latest
        synchronized payload per entity. Never includes password hashes,
        recovery-code hashes, session secrets or device-token hashes.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Account export
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/sync/merge:
    post:
      tags: [Sync]
      operationId: mergeLocalSyncData
      summary: Merge local data using normal sync push semantics
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncPushRequest'
      responses:
        '200':
          description: Operations committed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPushResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sync/snapshot:
    get:
      tags: [Sync]
      operationId: getSyncSnapshot
      summary: Get the current materialized sync state
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Full sync snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncSnapshot'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: koalacast_session
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    ProxyURL:
      name: url
      in: query
      required: true
      schema:
        type: string
        format: uri
        pattern: '^https?://'
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadGateway:
      description: Upstream resource could not be fetched
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    ChaptersResponse:
      type: object
      required: [chapters]
      properties:
        chapters:
          type: array
          items:
            $ref: '#/components/schemas/Chapter'
      additionalProperties: true
...
    TranscriptCuesResponse:
      type: object
      required: [cues]
      additionalProperties: false
      properties:
        cues:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/TranscriptCue'
...
    AddFeedRequest:
      type: object
      required: [feed_url]
      additionalProperties: true
      properties:
        feed_url:
          type: string
          format: uri
          pattern: '^https?://'
...